move : fetchUserProfile 파일 이동
