+++ src/app/api/fetchUserProfile.ts
